TIP-576: move PimTestCase to integration folder
