+++ composer.json
@@ -12,11 +12,11 @@
         "psr-0": {
             "": "src/",
             "Context": "features/",
-            "Pim\\Behat": "features/",
-            "Pim\\Integration": "features/"
+            "Pim\\Behat": "features/"
         },
         "psr-4": {
-            "Pim\\Upgrade\\": "upgrades/"
+            "Pim\\Upgrade\\": "upgrades/",
+            "Test\\Integration\\": "tests/"
         }
     },
     "require": {
