+++ src/Pim/Component/Catalog/tests/integration/Normalizer/Standard/AttributeGroupIntegration.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Pim\Component\Catalog\Tests\Integration\StandardFormat;
+namespace tests\integration\Pim\Component\Catalog\Standard;
 
-use Pim\Integration\PimTestCase;
+use Test\Integration\TestCase;
 
 /**
  * @author    Marie Bochu <marie.bochu@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AttributeGroupIntegration extends PimTestCase
+class AttributeGroupIntegration extends TestCase
 {
     protected $purgeDatabaseForEachTest = false;
