+++ src/Pim/Component/Catalog/tests/integration/Normalizer/Standard/CurrencyIntegration.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace Pim\Component\Catalog\Tests\Integration\StandardFormat;
+namespace tests\integration\Pim\Component\Catalog\Standard;
 
-use Pim\Integration\PimTestCase;
+use Test\Integration\TestCase;
 
 /**
  * @author    Marie Bochu <marie.bochu@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class LocaleIntegration extends PimTestCase
+class CurrencyIntegration extends TestCase
 {
-    public function testLocale()
+    public function testCurrency()
     {
         $expected = [
-            'code'    => 'en_US',
+            'code'    => 'USD',
             'enabled' => true,
         ];
 
-        $repository = $this->get('pim_catalog.repository.locale');
+        $repository = $this->get('pim_catalog.repository.currency');
         $serializer = $this->get('pim_serializer');
 
-        $result = $serializer->normalize($repository->findOneByIdentifier('en_US'), 'standard');
+        $result = $serializer->normalize($repository->findOneByIdentifier('USD'), 'standard');
 
         $this->assertSame($expected, $result);
     }
