+++ src/Pim/Component/Catalog/tests/integration/Normalizer/Standard/common.sql
