+++ src/Pim/Component/Catalog/tests/integration/Normalizer/Standard/products_mongodb.json
