+++ tests/catalog/technical/association_types.csv
