+++ tests/catalog/technical/attribute_groups.csv
