+++ tests/catalog/technical/attributes.csv
