+++ tests/catalog/technical/currencies.csv
