+++ tests/catalog/technical/group_types.csv
