+++ tests/catalog/technical/locales.csv
