+++ tests/catalog/technical/variant_groups.csv
